requirements.txt
absl-py==0.9.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
astor==0.8.1
astroid==2.4.2
attrs==20.1.0
Automat==20.2.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
bce-python-sdk==0.8.59
beautifulsoup4==4.6.0
bitarray==1.5.3
boto==2.49.0
boto3==1.14.23
botocore==1.17.23
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
distlib==0.3.1
Django==3.0.8
django-crontab==0.7.1
docker==4.3.1
docopt==0.6.2
docutils==0.15.2
elasticsearch==7.6.0
english==2020.7.0
fasttext==0.9.1
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Babel==2.0.0
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
gensim==3.8.1
graphviz==0.16
grpcio==1.30.0
h5py==2.10.0
happybase==1.2.0
hdfs==2.5.7
html2text==2020.1.16
hyperlink==20.0.1
identify==2.2.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==5.1.2
incremental==17.5.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
isodate==0.6.0
isort==5.5.3
itemadapter==0.1.0
itemloaders==1.0.2
itsdangerous==1.1.0
jedi==0.10.2
jieba==0.37
Jinja2==2.11.3
jmespath==0.10.0
joblib==0.16.0
jupyter-client==6.1.12
jupyter-core==4.6.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
ltp==4.1.3.post1
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
mock==4.0.2
monotonic==1.5
munch==2.5.0
neo4j-driver==1.7.6
neobolt==1.7.17
neotime==1.7.4
networkx==2.4
nltk==3.3
nodeenv==1.5.0
numpy
objgraph==3.5.0
openccpy==0.0.4.3
opencv-python==4.5.1.48
packaging==20.4
paddle-ernie==0.1.0.dev1
paddlepaddle-gpu==2.0.1
pandas
pansi==2020.7.3
parsel==1.6.0
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
ply==3.11
pre-commit==2.11.1
prettytable==2.1.0
prompt-toolkit
Protego==0.1.16
protobuf==3.13.0
ptyprocess==0.6.0
py2neo==2020.0.0
pyahocorasick==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pybloom-live==3.0.0
pycodestyle==2.6.0
pycparser
pycryptodome==3.10.1
pycurl==7.43.0.6
PyDispatcher==2.0.5
pyflakes==2.2.0
Pygments
pygtrie==2.4.2
PyHamcrest==2.0.2
PyHDFS==0.3.1
pykerberos==1.2.1
pylint==2.6.0
pyltp==0.2.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc==0.10.0
pytorch-crf==0.7.2
pytorch-ignite==0.2.0
pytorch-transformers
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
queuelib==1.5.0
rarfile==4.0
rdflib==5.0.0
regex==2020.7.14
requests
requests-kerberos
rope==0.17.0
s3transfer==0.3.3
sacremoses
scikit-learn==0.20.4
scipy==1.5.1
Scrapy==2.3.0
sentencepiece==0.1.91
service-identity==18.1.0
shellcheck-py==0.7.1.1
simplejson==3.17.2
six
smart-open==2.1.0
snownlp==0.12.3
socks==0
SPARQLWrapper==1.8.5
sqlparse==0.3.1
tensorboard==1.12.2
tensorflow==1.12.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
textrank4zh==0.3
threadpoolctl==2.1.0
thrift==0.13.0
thriftpy2==0.4.11
tokenizers==0.9.4
toml==0.10.1
torch==1.7.1
TorchCRF==1.0.4
torchvision
tornado==6.0.4
tqdm
traitlets==4.3.3
transformers==4.2.1
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal
urllib3
virtualenv==20.4.3
visualdl==2.1.1
w3lib==1.22.0
wcwidth
websocket-client==0.57.0
Werkzeug==1.0.1
Wikipedia-API==0.5.4
wikitools==1.3
wptools==0.4.17
wrapt==1.12.1
yapf==0.30.0
zipp==3.1.0
zope.interface==5.1.0
